:recycle: fix commonjs to es6
